Added Functionality to seperate expenses of multiple users

+++ controllers/expense.js
@@ -4,17 +4,20 @@ exports.addExpense = (req, res) => {
     const amount = req.body.amount;
     const description = req.body.description;
     const category = req.body.category;
+    const id = req.user.id;
     Expense.create({
         amount: amount,
         description: description,
-        category: category
+        category: category,
+        userId: id
     })
         .then(result => res.json(result))
         .catch(err => console.log(err));
 }
 
 exports.getExpense = (req, res) => {
-    Expense.findAll()
+    const id = req.user.id;
+    Expense.findAll({ where: { userId: id } })
         .then((expenses) => {
             res.json(expenses);
         })
@@ -23,7 +26,8 @@ exports.getExpense = (req, res) => {
 
 exports.deleteExpense = (req, res) => {
     const id = req.params.id;
-    Expense.destroy({ where: { id: id } })
+    const userId = req.user.id;
+    Expense.destroy({ where: { id: id, userId: userId } })
         .then(result => res.json(result))
         .catch(err => console.log(err));
 }
